requirements.txt
pytest==7.4.2
selenium==4.12.0
webdriver-manager==4.0.0
faker==17.0.0
allure-pytest==2.13.2
allure-behave==2.13.2
allure-python-commons-test==2.13.2
allure-python-commons==2.13.2

